chore(ui): rename test to tests
