+++ ui/design/src/components/Avatar/__tests__/index.test.tsx
