+++ ui/design/src/components/ProfileCard/__tests__/profile-card.test.tsx
@@ -2,7 +2,7 @@ import { cleanup, fireEvent, waitForElement } from '@testing-library/react';
 import * as React from 'react';
 import { act, create } from 'react-test-renderer';
 import { customRender, wrapWithTheme } from '../../../test-utils';
-import ProfileCard from '../../ProfileCard';
+import ProfileCard from '..';
 
 const mockProfileData = {
   ethAddress: '0x003410490050000320006570034567114572000',
