+++ ui/design/src/components/Spinner/__tests__/index.test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { create } from 'react-test-renderer';
-import Spinner from '../';
+import Spinner from '..';
 import { wrapWithTheme } from '../../../test-utils';
 
 describe('Spinner component', () => {
